add admin sign-up page, refactor login redirects, and implement user login with OTP verification
