+++ src/routes/auth/admin/sign-up/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+    import type { PageData } from './$types';
+
+    let { data }: { data: PageData } = $props();
+</script>
