+++ src/routes/auth/user/sign-up/parking-manager/+page.svelte
